# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Termix API",
    "version": "1.0.0",
    "description": "Comprehensive API for Termix SSH management, file operations, tunneling, and server monitoring. This API provides endpoints for managing SSH hosts, file operations, tunnel connections, server monitoring, user management, and system alerts.",
    "contact": {
      "name": "Termix Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Main database and authentication server"
    },
    {
      "url": "http://localhost:8083",
      "description": "SSH tunnel management server"
    },
    {
      "url": "http://localhost:8084",
      "description": "SSH file manager server"
    },
    {
      "url": "http://localhost:8085",
      "description": "Server statistics and monitoring server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "parameters": {
...
      "userId": {
        "name": "userId",
        "in": "path",
        "description": "The user identifier",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
  },
  "paths": {
...
    "/alerts/user/{userId}": {
      "get": {
        "summary": "Get alerts for specific user",
        "description": "Get alerts for specific user. **Server: localhost:8081**",
        "operationId": "getUserAlerts",
        "tags": ["Alerts"],
        "parameters": [
          {
            "$ref": "#/components/parameters/userId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "alerts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "title": { "type": "string" },
                          "message": { "type": "string" },
                          "expiresAt": { "type": "string" },
                          "priority": {
                            "type": "string",
                            "enum": ["low", "medium", "high", "critical"]
                          },
                          "type": {
                            "type": "string",
                            "enum": ["info", "warning", "error", "success"]
                          },
                          "actionUrl": { "type": "string" },
                          "actionText": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/alerts/dismiss": {
      "post": {
        "summary": "Dismiss an alert",
        "description": "Dismiss an alert. **Server: localhost:8081**",
        "operationId": "dismissAlert",
        "tags": ["Alerts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string" },
                  "alertId": { "type": "string" }
                },
                "required": ["userId", "alertId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}